Extend soaper-dl's functionality.
This adds;
The ability to set a specific directory with -x. (does not have a default due to not knowing how to do that, will fix later.)
Fixes an error with movies/shows that has it not download due to having a colon in it's name.
Moved all temporary files to /tmp/soaper-dl (folder still needs to be created every time the script is ran, will fix later.)
Fixes a spelling mistake in select_episodes_to_download function.
And probably much more that I've forgotten.